DodgeFinal
Added health to the player, which loses health every time it touches the mods. Apples will randomly appear on the map, and one score will be added if you pick them up. Pressing the space key will give the player a protective cover against damage from mods, but player will not be able to pick up apples, and the protective cover has a ten-second cooldown